+++ modules/App/GraphQL/Types/FieldTypes.php
@@ -29,6 +29,11 @@ class FieldTypes {
         foreach ($meta['fields'] as $field) {
 
             $def = self::getType($field);
+            $name = $field['name'];
+
+            if (is_numeric($name)) {
+                $name = "_{$name}_";
+            }
 
             if ($def) {
 
@@ -37,9 +42,9 @@ class FieldTypes {
                 }
 
                 if ($field['multiple']) {
-                    $fields[$field['name']] = Type::listOf($def['type']);
+                    $fields[$name] = Type::listOf($def['type']);
                 } else {
-                    $fields[$field['name']] = $def;
+                    $fields[$name] = $def;
                 }
             }
         }