+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Update vendor libs
 - Fix field-object js error
 - Don't apply default content fields on update (saveItem)
+- Fix GraphQL error when numeric field names are used
 
 
 ## 2.2.2 (2022-08-20)